Finish ignition solution for update scripts.
