+++ src/Exceptions/ComposerJsonMissingPreUpdateCmdException.php
@@ -5,12 +5,12 @@ namespace Statamic\Exceptions;
 use Exception;
 use Facade\IgnitionContracts\ProvidesSolution;
 use Facade\IgnitionContracts\Solution;
-use Statamic\Ignition\Solutions\AddComposerPreUpdateCmd;
+use Statamic\Ignition\Solutions\EnableComposerUpdateScripts;
 
 class ComposerJsonMissingPreUpdateCmdException extends Exception implements ProvidesSolution
 {
     public function getSolution(): Solution
     {
-        return new AddComposerPreUpdateCmd;
+        return new EnableComposerUpdateScripts;
     }
 }
